refactor(mounts): split the 1700-line Mounts.tsx into views/mounts/
Pure move, no behavior change: the page kept the whole feature in one file —
the mount list, the add form, four setup panels, a stepper and the provider
picker — and every one of them was reached by scrolling past the others.

  views/Mounts.tsx        the page: load, poll, restart, modal host
  views/mounts/MountList  UploadQueue + MountRow
  views/mounts/AddMount   the add form
  views/mounts/setup      the six setup flows + the provider picker
  views/mounts/links      the pure decisions (already tested)

Route and rendered output are identical.
